Finished Camera!
The 3rd person camera completely works now (minus it not clipping through objects but that wasn't planned to be done right now), and pressing escape toggles the mouse being visible and being locked to the game screen.
Also started adding animations, the triggers and the bool for walking, start, and end cycles have been added in the input stuff and theoretically should function fine, but need to tweak the exit time and parameters of the actual animations in the chart to not be stuttering on loops and transitions.
Big thing to add in the future is either a double jump or dash (or both), and adding the levels should be soon after.
Build isn't current, but pushing these changes before building to be safe.