Character Added!
Added a free unity 3d model to be the character I'm using. I did the animations for it since the given animations don't work, but the ones I made look pretty good! Still need to make the jumping and falling animations, and I need to make some levels.
Might want to add wall climbing/jumping mechanics, likely going to add a dash that can be used midair and on the ground. Almost certainly going to add an attack with the wand the 3d model came with. Big hat in time inspiration lol.