Enemies and QOL Changes added
Added enemies, that currently do not do anything nor are able to get attacked, but if touched will respawn you so they are basically a spike for now LOL.
The real challenge in this one was learning how to change the material of the slimes in the code, as they were in a skinned mesh renderer and I haven't messed around with 3d that much before. Solved that by just getting the array of materials, editing the material I know will be the body, then setting the whole array again. Otherwise, the ScriptableObject of the enemy data can be called to randomly spawn a random enemy with a random color at any given location.
Minor QOL I did was to add camera scrolling so its easier to see some of the smaller things, Added some blur over the wand's attacks to make it more clear where and when the player can hit the enemy, and other minor tweaks I can't remember off the top of my head.